add potential Post module.
Posts are currently listed under /wiki.
Specific a post can be viewed under /wiki/[slug] where slug is the
defined slug of the post object.
Multiple users can be related to one post, and queried using
user.created_posts or user.edited_posts. Detection takes place in
pages.forms.PageForm.save method.

Additionally I don't plan to abandon the Post object, it will be
converted later to a discussion board, in a forum style in which people
are allowed to post things without the strict restriction to a specific
topic/namespace or however you want to call it.

Current implementation of Page does not hold Tag, I'd like to discuss
importance of a tag system further and opened #17.